requirements_clean.txt
# Core FastAPI
fastapi==0.135.1
uvicorn==0.42.0
python-multipart==0.0.22
python-dotenv==1.2.2

# Document Processing
PyMuPDF==1.27.2
python-pptx==1.0.2
easyocr==1.7.2
opencv-python-headless==4.13.0.92
Pillow==12.0.0

# ML & Embeddings
sentence-transformers==5.3.0
torch==2.5.1
transformers==5.3.0
faiss-cpu==1.13.2
numpy==2.3.5

# NLP & Text Processing
beautifulsoup4==4.14.3
lxml==6.0.2
tiktoken==0.12.0

# Data & Config
pandas==2.3.3
PyYAML==6.0.3
pydantic==2.12.5

# Utilities
requests==2.32.5
tqdm==4.67.3
Jinja2==3.1.3
